Moved placement of sharing component
