+++ components/resourceDetails/detailsLoading/index.jsx
@@ -23,34 +23,21 @@
 
 import React from "react";
 import {
-  Typography,
   Divider,
-  List,
-  ListItem,
-  ListItemIcon,
-  ListItemText,
   createStyles,
+  Accordion,
+  AccordionSummary,
+  AccordionDetails,
 } from "@material-ui/core";
 import { makeStyles } from "@material-ui/styles";
 import T from "prop-types";
-import ShareIcon from "@material-ui/icons/Share";
-import DeleteIcon from "@material-ui/icons/Delete";
 import Skeleton from "@material-ui/lab/Skeleton";
-import { DETAILS_CONTEXT_ACTIONS } from "../../../src/contexts/detailsMenuContext";
-import ResourceLink from "../../resourceLink";
-import DeleteLink from "../deleteLink";
+import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
+import { ActionMenu, ActionMenuItem } from "@inrupt/prism-react-components";
 import styles from "./styles";
+import DownloadLink from "../../downloadLink";
 
 const useStyles = makeStyles((theme) => createStyles(styles(theme)));
-
-/* eslint react/jsx-props-no-spreading: 0 */
-const SharingLink = React.forwardRef((linkProps, ref) => (
-  <ResourceLink
-    {...linkProps}
-    action={DETAILS_CONTEXT_ACTIONS.SHARING}
-    ref={ref}
-  />
-));
 
 function DetailsLoading({ name, iri, onDelete, onDeleteError }) {
   const classes = useStyles();
@@ -65,49 +52,17 @@ function DetailsLoading({ name, iri, onDelete, onDeleteError }) {
 
       <Divider />
 
-      <section className={classes.centeredSection}>
-        <h5 className={classes["content-h5"]}>Actions</h5>
-        <List>
-          <ListItem button component={SharingLink} resourceIri={iri}>
-            <ListItemIcon>
-              <ShareIcon />
-            </ListItemIcon>
-            <ListItemText primary="Sharing &amp; App Permissions" />
-          </ListItem>
+      <Accordion disabled>
+        <AccordionSummary>Actions</AccordionSummary>
+      </Accordion>
 
-          <ListItem
-            button
-            component={DeleteLink}
-            resourceIri={iri}
-            onDelete={onDelete}
-            onDeleteError={onDeleteError}
-          >
-            <ListItemIcon>
-              <DeleteIcon />
-            </ListItemIcon>
-            <ListItemText primary="Delete" />
-          </ListItem>
-        </List>
-      </section>
+      <Accordion disabled>
+        <AccordionSummary>Details</AccordionSummary>
+      </Accordion>
 
-      <Divider />
-
-      <section className={classes.centeredSection}>
-        <h5 className={classes["content-h5"]}>Details</h5>
-      </section>
-
-      <Divider />
-
-      <section className={classes.centeredSection}>
-        <List>
-          <ListItem className={classes.listItem}>
-            <Typography className={classes.detailText}>Thing Type:</Typography>
-            <Skeleton width={150} />
-          </ListItem>
-        </List>
-      </section>
-
-      <Divider />
+      <Accordion disabled>
+        <AccordionSummary>Permissions</AccordionSummary>
+      </Accordion>
     </>
   );
 }
