+++ components/resourceDetails/index.jsx
@@ -19,30 +19,30 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import React from "react";
+import React, { useState } from "react";
 import {
-  Button,
+  Accordion,
+  AccordionDetails,
+  AccordionSummary,
   createStyles,
   Divider,
   List,
   ListItem,
-  ListItemIcon,
-  ListItemText,
   Typography,
 } from "@material-ui/core";
-import ShareIcon from "@material-ui/icons/Share";
-import DeleteIcon from "@material-ui/icons/Delete";
 import { makeStyles } from "@material-ui/styles";
 import T from "prop-types";
 import { useSession } from "@inrupt/solid-ui-react";
+import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
+import { ActionMenu, ActionMenuItem } from "@inrupt/prism-react-components";
 import ResourceLink from "../resourceLink";
 import styles from "./styles";
-import { parseUrl } from "../../src/stringHelpers";
-import { DETAILS_CONTEXT_ACTIONS } from "../../src/contexts/detailsMenuContext";
-import DeleteLink from "./deleteLink";
+import DeleteLink from "../deleteLink";
+import DownloadLink from "../downloadLink";
+import ResourceSharing from "./resourceSharing";
 
-const TESTCAFE_ID_SHARE_PERMISSIONS_BUTTON = "share-permissions-button";
 const TESTCAFE_ID_DOWNLOAD_BUTTON = "download-resource-button";
+const TESTCAFE_ID_SHARE_PERMISSIONS_BUTTON = "share-permissions-button";
 const TESTCAFE_ID_TITLE = "resource-title";
 
 export function displayType(types) {
@@ -53,75 +53,21 @@ export function displayType(types) {
 
 const useStyles = makeStyles((theme) => createStyles(styles(theme)));
 
-export function forceDownload(name, file) {
-  const a = document.createElement("a");
-  a.style.display = "none";
-  a.href = window.URL.createObjectURL(file);
-  a.setAttribute("download", name);
-
-  document.body.appendChild(a);
-
-  a.click();
-
-  window.URL.revokeObjectURL(a.href);
-  document.body.removeChild(a);
-}
-
-export function downloadResource(iri, fetch) {
-  return () => {
-    const { pathname } = parseUrl(iri);
-    const name = pathname.replace(/\//g, "");
-
-    fetch(iri)
-      .then((response) => response.blob())
-      .then((file) => forceDownload(name, file))
-      .catch((e) => e);
-  };
-}
-
-export function DownloadLink(props) {
-  const { session } = useSession();
-  const { type, iri, className } = props;
-
-  if (type.match(/container/i)) return null;
-
-  return (
-    <Button
-      data-testid={TESTCAFE_ID_DOWNLOAD_BUTTON}
-      variant="contained"
-      onClick={downloadResource(iri, session.fetch)}
-      className={className}
-    >
-      Download
-    </Button>
-  );
-}
-
-DownloadLink.propTypes = {
-  type: T.string.isRequired,
-  iri: T.string.isRequired,
-  className: T.string,
-};
-
-DownloadLink.defaultProps = {
-  className: null,
-};
-
-/* eslint react/jsx-props-no-spreading: 0 */
-const SharingLink = React.forwardRef((linkProps, ref) => (
-  <ResourceLink
-    {...linkProps}
-    data-testid={TESTCAFE_ID_SHARE_PERMISSIONS_BUTTON}
-    action={DETAILS_CONTEXT_ACTIONS.SHARING}
-    ref={ref}
-  />
-));
-
 export default function ResourceDetails({ resource, onDelete, onDeleteError }) {
+  const [sharingExpanded, setSharingExpanded] = useState(false);
   const classes = useStyles();
-  const { iri, name, types } = resource;
+  const {
+    iri,
+    name,
+    types,
+    permissions,
+    defaultPermissions,
+    dataset,
+  } = resource;
   const type = displayType(types);
+  const actionMenuBem = ActionMenu.useBem();
 
+  const expandIcon = <ExpandMoreIcon />;
   return (
     <>
       <section className={classes.headerSection}>
@@ -136,62 +82,82 @@ export default function ResourceDetails({ resource, onDelete, onDeleteError }) {
 
       <Divider />
 
-      <section className={classes.centeredSection}>
-        <h5 className={classes["content-h5"]}>Actions</h5>
-        <List>
-          <ListItem button component={SharingLink} resourceIri={iri}>
-            <ListItemIcon>
-              <ShareIcon />
-            </ListItemIcon>
-            <ListItemText primary="Sharing &amp; App Permissions" />
-          </ListItem>
+      <Accordion defaultExpanded>
+        <AccordionSummary expandIcon={expandIcon}>Actions</AccordionSummary>
+        <AccordionDetails>
+          <ActionMenu>
+            <ActionMenuItem>
+              <DownloadLink
+                className={actionMenuBem("action-menu__trigger")}
+                data-testid={TESTCAFE_ID_DOWNLOAD_BUTTON}
+                iri={iri}
+                type={type}
+              >
+                Download
+              </DownloadLink>
+            </ActionMenuItem>
+            <ActionMenuItem>
+              <button
+                className={actionMenuBem("action-menu__trigger")}
+                data-testid={TESTCAFE_ID_SHARE_PERMISSIONS_BUTTON}
+                type="button"
+                onClick={() => setSharingExpanded(true)}
+              >
+                Sharing & App Permissions
+              </button>
+            </ActionMenuItem>
+            <ActionMenuItem>
+              <DeleteLink
+                className={actionMenuBem("action-menu__trigger", "danger")}
+                resourceIri={iri}
+                name={name}
+                onDelete={onDelete}
+                onDeleteError={onDeleteError}
+                data-testid={TESTCAFE_ID_DOWNLOAD_BUTTON}
+              >
+                Delete
+              </DeleteLink>
+            </ActionMenuItem>
+          </ActionMenu>
+        </AccordionDetails>
+      </Accordion>
 
-          <ListItem
-            button
-            component={DeleteLink}
-            resourceIri={iri}
-            name={name}
-            onDelete={onDelete}
-            onDeleteError={onDeleteError}
-          >
-            <ListItemIcon>
-              <DeleteIcon />
-            </ListItemIcon>
-            <ListItemText primary="Delete" />
-          </ListItem>
-        </List>
-      </section>
+      <Accordion>
+        <AccordionSummary expandIcon={expandIcon}>Details</AccordionSummary>
+        <AccordionDetails>
+          <section className={classes.centeredSection}>
+            <List>
+              <ListItem className={classes.listItem}>
+                <Typography className={classes.detailText}>
+                  Thing Type:
+                </Typography>
+                <Typography
+                  className={`${classes.typeValue} ${classes.detailText}`}
+                >
+                  {type}
+                </Typography>
+              </ListItem>
+            </List>
+          </section>
+        </AccordionDetails>
+      </Accordion>
 
-      <Divider />
-
-      <section className={classes.centeredSection}>
-        <h5 className={classes["content-h5"]}>Details</h5>
-      </section>
-
-      <Divider />
-
-      <section className={classes.centeredSection}>
-        <List>
-          <ListItem className={classes.listItem}>
-            <Typography className={classes.detailText}>Thing Type:</Typography>
-            <Typography
-              className={`${classes.typeValue} ${classes.detailText}`}
-            >
-              {type}
-            </Typography>
-          </ListItem>
-        </List>
-      </section>
-
-      <Divider />
-
-      <section className={classes.centeredSection}>
-        <DownloadLink
-          type={type}
-          iri={iri}
-          className={classes.downloadButton}
-        />
-      </section>
+      <Accordion
+        expanded={sharingExpanded}
+        onChange={() => setSharingExpanded(!sharingExpanded)}
+      >
+        <AccordionSummary expandIcon={expandIcon}>Permissions</AccordionSummary>
+        <AccordionDetails>
+          <div>
+            <ResourceSharing
+              iri={iri}
+              permissions={permissions}
+              defaultPermissions={defaultPermissions}
+              dataset={dataset}
+            />
+          </div>
+        </AccordionDetails>
+      </Accordion>
     </>
   );
 }
@@ -201,6 +167,10 @@ ResourceDetails.propTypes = {
     iri: T.string.isRequired,
     name: T.string.isRequired,
     types: T.arrayOf(T.string).isRequired,
+    permissions: T.arrayOf(T.object).isRequired,
+    defaultPermissions: T.arrayOf(T.object).isRequired,
+    // eslint-disable-next-line react/forbid-prop-types
+    dataset: T.object.isRequired,
   }),
   onDelete: T.func,
   onDeleteError: T.func,
@@ -211,6 +181,9 @@ ResourceDetails.defaultProps = {
     iri: "",
     name: "",
     types: [],
+    permissions: [],
+    defaultPermissions: [],
+    dataset: null,
   },
   onDelete: () => {},
   onDeleteError: () => {},
