+++ components/resourceDetails/index.test.jsx
@@ -20,19 +20,11 @@
  */
 
 import React from "react";
-import { shallow } from "enzyme";
-import { shallowToJson } from "enzyme-to-json";
 import * as Router from "next/router";
 import { mountToJson } from "../../__testUtils/mountWithTheme";
-import * as stringHelpers from "../../src/stringHelpers";
 import ResourceDetails, * as resourceDetailFns from "./index";
 
-const {
-  displayType,
-  DownloadLink,
-  downloadResource,
-  forceDownload,
-} = resourceDetailFns;
+const { displayType } = resourceDetailFns;
 
 jest.mock("../../src/hooks/solidClient");
 
@@ -106,80 +98,5 @@ describe("displayType", () => {
   test("it returns the first type if types", () => {
     const types = ["A Type"];
     expect(displayType(types)).toEqual(types[0]);
-  });
-});
-
-describe("forceDownload", () => {
-  test("it creates an anchor with an object url and clicks it", () => {
-    const setAttributeMock = jest.fn();
-    const clickMock = jest.fn();
-    const revokeObjectURLMock = jest.fn();
-    const appendChildMock = jest.fn();
-    const removeChildMock = jest.fn();
-    const mockAnchor = {
-      style: { display: "block" },
-      setAttribute: setAttributeMock,
-      click: clickMock,
-    };
-    const file = new Blob(["file"]);
-
-    window.URL = {
-      createObjectURL: () => "object-url",
-      revokeObjectURL: revokeObjectURLMock,
-    };
-
-    jest.spyOn(document, "createElement").mockReturnValue(mockAnchor);
-    jest
-      .spyOn(document.body, "appendChild")
-      .mockImplementationOnce(appendChildMock);
-    jest
-      .spyOn(document.body, "removeChild")
-      .mockImplementationOnce(removeChildMock);
-
-    forceDownload("filename", file);
-
-    expect(mockAnchor.style.display).toEqual("none");
-    expect(mockAnchor.href).toEqual("object-url");
-    expect(setAttributeMock).toHaveBeenCalledWith("download", "filename");
-    expect(appendChildMock).toHaveBeenCalledWith(mockAnchor);
-    expect(clickMock).toHaveBeenCalled();
-    expect(revokeObjectURLMock).toHaveBeenCalledWith("object-url");
-    expect(removeChildMock).toHaveBeenCalledWith(mockAnchor);
-  });
-});
-
-describe("downloadResource", () => {
-  test("it returns a handler to download the resource", () => {
-    const iri = "http://example.com/resource";
-    const blobMock = jest.fn();
-    const responseMock = { blob: blobMock };
-    const fetchMock = jest.fn().mockResolvedValue(responseMock);
-    const handler = downloadResource(iri, fetchMock);
-
-    blobMock.mockResolvedValue("file");
-
-    jest
-      .spyOn(stringHelpers, "parseUrl")
-      .mockReturnValue({ pathname: "/resource" });
-
-    handler();
-
-    expect(fetchMock).toHaveBeenCalledWith(iri);
-  });
-});
-
-describe("DownloadLink", () => {
-  test("returns null if resource is a container", () => {
-    const type = "container";
-    const iri = "http://example.com/resource";
-    const tree = shallow(<DownloadLink type={type} iri={iri} />);
-    expect(shallowToJson(tree)).toMatchSnapshot();
-  });
-
-  test("returns a download button if resource is not a container", () => {
-    const type = "foo";
-    const iri = "http://example.com/resource";
-    const tree = shallow(<DownloadLink type={type} iri={iri} />);
-    expect(shallowToJson(tree)).toMatchSnapshot();
   });
 });
